prevent base64 wrapping at 76 chars, fix logic bug in deployment and switch to LTS for dotnet install
base64 wraps at char 76 and introduces a white space in the encoded value, assuming it's long enough. This, while rare, will result in an: unmatched "/" error when sed is used to replace the auth value. It then breaks pulling from ACR as the ACR creds are missing the auth (it remained as $auth). This will disable wrapping, thus eliminating that possibility. Fix bug in logic for Profisee deployment, switch to LTS for dotnet install